backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo==4.6.2
motor==3.3.2
pydantic==2.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
firebase-admin==6.4.0
python-dotenv==1.0.1 